fixed truncation value argument and used np.clip

+++ generate_lighting.py
@@ -64,7 +64,7 @@ if __name__ == '__main__':
     parser.add_argument("--n_samples", help="Number of images to generate",
                         type=int, default=100)
     parser.add_argument("--truncation", help="Truncation of latent styles",
-                        type=int, default=0.7)
+                        type=float, default=0.7)
     parser.add_argument("--batch_size", help="Number of images per mini batch",
                         type=int, default=16)
     args = parser.parse_args()
