+++ generate_poses.py
@@ -75,7 +75,7 @@ if __name__ == '__main__':
     parser.add_argument("--type", help="The path to aligned face images",
                         type=str, default="yaw")
     parser.add_argument("--truncation", help="Truncation of latent styles",
-                        type=int, default=0.7)
+                        type=float, default=0.7)
     parser.add_argument("--n_samples", help="Number of images to generate",
                         type=int, default=100)
     parser.add_argument("--batch_size", help="Number of images per mini batch",
